pylintrc: re-enable too-many-lines, part 2
To get under the 1000 line limit, move the ceph
exception classes to more specific files. Updated
tests accordingly.
