+++ hotsos/defs/tests/scenarios/storage/ceph/ceph-mon/laggy_pgs.yaml
@@ -11,7 +11,7 @@ data-root:
     - sos_commands/systemd/systemctl_list-unit-files
 mock:
   patch:
-    hotsos.core.plugins.storage.ceph.CephCluster.require_osd_release:
+    hotsos.core.plugins.storage.ceph_cluster.CephCluster.require_osd_release:
       kwargs:
         new: octopus
 raised-issues:
