+++ hotsos/defs/tests/scenarios/storage/ceph/ceph-mon/pg_imbalance.yaml
@@ -10,7 +10,7 @@ data-root:
     - sos_commands/systemd/systemctl_list-unit-files
 mock:
   patch:
-    hotsos.core.plugins.storage.ceph.CephCrushMap.ceph_report:
+    hotsos.core.plugins.storage.ceph_cluster.CephCrushMap.ceph_report:
       kwargs:
         new:
           osdmap:
@@ -19,7 +19,7 @@ mock:
               - pg_autoscale_mode: 'off'
               - pool: 2
               - pg_autoscale_mode: 'on'
-    hotsos.core.plugins.storage.ceph.CephCluster.cluster_has_non_empty_pools:
+    hotsos.core.plugins.storage.ceph_cluster.CephCluster.cluster_has_non_empty_pools:
       kwargs:
         new: true
 raised-issues:
