+++ hotsos/defs/tests/scenarios/storage/ceph/ceph-mon/pg_overdose.yaml
@@ -20,7 +20,7 @@ data-root:
     - sos_commands/systemd/systemctl_list-unit-files
 mock:
   patch:
-    hotsos.core.plugins.storage.ceph.CephCluster.health_status:
+    hotsos.core.plugins.storage.ceph_cluster.CephCluster.health_status:
       kwargs:
         new: HEALTH_WARN
 raised-issues:
