+++ hotsos/defs/tests/scenarios/storage/ceph/ceph-mon/required_osd_release_mismatch.yaml
@@ -23,7 +23,7 @@ data-root:
     - sos_commands/systemd/systemctl_list-unit-files
 mock:
   patch:
-    hotsos.core.plugins.storage.ceph.CephCluster.require_osd_release:
+    hotsos.core.plugins.storage.ceph_cluster.CephCluster.require_osd_release:
       kwargs:
         new: octopus
 raised-issues:
